fix(@crates/eur-vision): build issues on windows by switching to 0.6 of xcap

+++ crates/app/eur-vision/src/lib.rs
@@ -76,7 +76,7 @@ pub fn capture_monitor_region(
     // let region_height = height.min(monitor_height - y) as u32;
 
     let image_region = monitor
-        .capture_region(x as i32, y as i32, width, height)
+        .capture_region(x, y, width, height)
         .map_err(|e| anyhow!("Failed to capture region: {}", e))?;
 
     Ok(image_region)
@@ -96,7 +96,7 @@ pub fn capture_monitor_region_rgba(
     // let region_height = height.min(monitor_height - y) as u32;
 
     let image_region = monitor
-        .capture_region(x as i32, y as i32, width, height)
+        .capture_region(x, y, width, height)
         .map_err(|e| anyhow!("Failed to capture region: {}", e))?;
 
     Ok(image_region)
